requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
diffusers==0.28.2
filelock==3.14.0
fsspec==2024.6.0
huggingface-hub==0.23.2
idna==3.7
importlib-metadata==7.1.0
jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==24.0
pillow==10.3.0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
sympy==1.12.1
torch==2.1.0a0+41361538.nv23.6
tqdm==4.66.4
typing-extensions==4.12.1
urllib3==2.2.1
zipp==3.19.2
